requirements.txt
langchain==0.0.82
openai==0.26.4
python-dotenv==0.21.1
streamlit==1.12.0
youtube_transcript_api==0.5.0
